+++ README.md
@@ -1,7 +1,12 @@
-# Daohaus Moloch V2 Factory Graph
+# Moloch V2 Factory Graph for multi summoner DAOs
 
 wip: moloch events.
 
 built by:
+LexDAO 
+
+forked from and thanks to DAOHaus Supergraph built by: 
 Welcome to FlatLand:
 http://manifesto.designerdao.eth.link/
+ 
+
