+++ package.json
@@ -1,15 +1,14 @@
 {
-  "name": "daohaus-moloch-v2-factory",
+  "name": "lexdao-multisummoner-moloch-factory",
   "license": "UNLICENSED",
   "scripts": {
     "codegen": "graph codegen",
     "build": "graph build",
-    "deploy-main": "node scripts/deploy-prep.js mainnet && graph deploy --node https://api.thegraph.com/deploy/ --ipfs https://api.thegraph.com/ipfs/ odyssy-automaton/daohaus",
-    "deploy-kovan": "node scripts/deploy-prep.js kovan && graph deploy --node https://api.thegraph.com/deploy/ --ipfs https://api.thegraph.com/ipfs/ odyssy-automaton/daohaus-kovan",
-    "deploy-rinkeby": "node scripts/deploy-prep.js rinkeby && graph deploy --node https://api.thegraph.com/deploy/ --ipfs https://api.thegraph.com/ipfs/ odyssy-automaton/daohaus-rinkeby",
-    "create-local": "graph create --node http://localhost:8020/ odyssy-automaton/daohaus",
-    "remove-local": "graph remove --node http://localhost:8020/ odyssy-automaton/daohaus",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 odyssy-automaton/daohaus"
+    "deploy-main": "node scripts/deploy-prep.js mainnet && graph deploy --node https://api.thegraph.com/deploy/ --ipfs https://api.thegraph.com/ipfs/ wswarren12/lex-dao-multi-summoner-moloch",
+    "deploy-rinkeby": "node scripts/deploy-prep.js rinkeby && graph deploy --node https://api.thegraph.com/deploy/ --ipfs https://api.thegraph.com/ipfs/ lex-dao-multi-summoner-moloch-rinkeby",
+    "create-local": "graph create --node http://localhost:8020/ wswarren12/lexdaomoloch",
+    "remove-local": "graph remove --node http://localhost:8020/ wswarren12/lexdaomoloch",
+    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 wswarren12/lexdaomoloch"
   },
   "dependencies": {
     "@graphprotocol/graph-cli": "0.17.1",
