+++ subgraph.yaml
@@ -3,39 +3,11 @@ schema:
   file: ./schema.graphql
 dataSources:
   - kind: ethereum/contract
-    name: V1Factory
-    network: rinkeby
+    name: MolochSummoner
+    network: mainnet 
     source:
-      address: '0x610247467d0dfA8B477ad7Dd1644e86CB2a79F8F'
-      abi: V1Factory
-      startBlock: 6494343
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.3
-      language: wasm/assemblyscript
-      entities:
-        - Register
-      abis:
-        - name: V1Factory
-          file: ./abis/V1Factory.json
-        - name: V1Moloch
-          file: ./abis/V1Moloch.json
-        - name: Guildbank
-          file: ./abis/Guildbank.json
-        - name: Erc20
-          file: ./abis/Erc20.json
-        - name: Erc20Bytes32
-          file: ./abis/Erc20Bytes32.json
-      eventHandlers:
-        - event: 'Register(address,indexed address,uint256,string,uint256)'
-          handler: handleRegisterV1
-      file: ./src/factory-mapping.ts
-  - kind: ethereum/contract
-    name: V2Factory
-    network: rinkeby
-    source:
-      address: '0x763b61A62EF076ad960E1d513713B2aeD7C1b88e'
-      abi: V2Factory
+      address: '' # add address 
+      abi: MolochSummoner
       startBlock: 6494329
     mapping:
       kind: ethereum/events
@@ -44,70 +16,23 @@ dataSources:
       entities:
         - Register
       abis:
-        - name: V2Factory
-          file: ./abis/V2Factory.json
+        - name: MolochSummoner
+          file: ./abis/MolochSummoner.json
         - name: Erc20
           file: ./abis/Erc20.json
         - name: Erc20Bytes32
           file: ./abis/Erc20Bytes32.json
       eventHandlers:
         - event: >-
-            Register(uint256,address,indexed
-            address,address[],uint256,uint256,uint256,uint256,uint256,uint256,uint256,string,uint256)
-          handler: handleRegisterV2
-        - event: Delete(address)
-          handler: handleDeleteV2
+            Summoned(indexed address, indexed address)
+          handler: handleRegisterMoloch
       file: ./src/factory-mapping.ts
 templates:
   - kind: ethereum/contract
-    name: MolochV1Template
-    network: rinkeby
+    name: MolochTemplate
+    network: mainnet
     source:
-      abi: V1Moloch
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.3
-      language: wasm/assemblyscript
-      file: ./src/v1-mapping.ts
-      entities:
-        - Moloch
-        - Member
-        - Vote
-        - Proposal
-        - Badge
-      abis:
-        - name: V1Moloch
-          file: ./abis/V1Moloch.json
-        - name: Guildbank
-          file: ./abis/Guildbank.json
-        - name: Erc20
-          file: ./abis/Erc20.json
-        - name: Erc20Bytes32
-          file: ./abis/Erc20Bytes32.json
-      eventHandlers:
-        - event: 'SummonComplete(indexed address,uint256)'
-          handler: handleSummonComplete
-        - event: >-
-            SubmitProposal(uint256,indexed address,indexed address,indexed
-            address,uint256,uint256)
-          handler: handleSubmitProposal
-        - event: 'SubmitVote(indexed uint256,indexed address,indexed address,uint8)'
-          handler: handleSubmitVote
-        - event: >-
-            ProcessProposal(indexed uint256,indexed address,indexed
-            address,uint256,uint256,bool)
-          handler: handleProcessProposal
-        - event: 'Ragequit(indexed address,uint256)'
-          handler: handleRagequit
-        - event: 'Abort(indexed uint256,address)'
-          handler: handleAbort
-        - event: 'UpdateDelegateKey(indexed address,address)'
-          handler: handleUpdateDelegateKey
-  - kind: ethereum/contract
-    name: MolochV2Template
-    network: rinkeby
-    source:
-      abi: V2Moloch
+      abi: Moloch
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.3
@@ -122,8 +47,8 @@ templates:
         - Proposal
         - Badge
       abis:
-        - name: V2Moloch
-          file: ./abis/V2Moloch.json
+        - name: Moloch
+          file: ./abis/Moloch.json
         - name: Erc20
           file: ./abis/Erc20.json
         - name: Erc20Bytes32
